--- a/app/server/models/credit_transfer.py
+++ b/app/server/models/credit_transfer.py
@@ -49,6 +49,8 @@ class CreditTransfer(ManyOrgBase, BlockchainTaskableBase):
     token_id        = db.Column(db.Integer, db.ForeignKey(Token.id))
 
     sender_transfer_account_id       = db.Column(db.Integer, db.ForeignKey("transfer_account.id"))
+    sender_transfer_account          = db.relationship('TransferAccount', foreign_keys=[sender_transfer_account_id])
+
     recipient_transfer_account_id    = db.Column(db.Integer, db.ForeignKey("transfer_account.id"))
 
     sender_blockchain_address_id    = db.Column(db.Integer, db.ForeignKey("blockchain_address.id"))
@@ -190,7 +192,7 @@ class CreditTransfer(ManyOrgBase, BlockchainTaskableBase):
                  transfer_subtype: TransferSubTypeEnum=None,
                  transfer_mode: TransferModeEnum = None,
                  is_ghost_transfer=False):
-
+        print("INNIT!!!")
         if amount < 0:
             raise Exception("Negative amount provided")
         self.transfer_amount = amount