--- a/app/server/models/transfer_account.py
+++ b/app/server/models/transfer_account.py
@@ -62,9 +62,9 @@ class TransferAccount(OneOrgBase, ModelBase, SoftDelete):
         lazy='joined'
     )
 
-    credit_sends       = db.relationship('CreditTransfer', foreign_keys='CreditTransfer.sender_transfer_account_id')
+    credit_sends       = db.relationship('CreditTransfer', foreign_keys='CreditTransfer.sender_transfer_account_id', back_populates='sender_transfer_account')
 
-    credit_receives    = db.relationship('CreditTransfer', foreign_keys='CreditTransfer.recipient_transfer_account_id')
+    credit_receives    = db.relationship('CreditTransfer', foreign_keys='CreditTransfer.recipient_transfer_account_id', back_populates='recipient_transfer_account')
 
     spend_approvals_given = db.relationship('SpendApproval', backref='giving_transfer_account',
                                             foreign_keys='SpendApproval.giving_transfer_account_id')