--- a/app/server/models/utils.py
+++ b/app/server/models/utils.py
@@ -12,6 +12,7 @@ from sqlalchemy import or_
 import server
 from server import db, bt, AppQuery
 from server.exceptions import OrganisationNotProvidedException, ResourceAlreadyDeletedError
+from server.utils.transfer_enums import BlockchainStatus
 
 
 @contextmanager
@@ -200,19 +201,21 @@ class ModelBase(db.Model):
     updated = db.Column(db.DateTime, default=datetime.datetime.utcnow, onupdate=datetime.datetime.utcnow)
 
 
+from server.models.worker_messages import WorkerMessages
 class BlockchainTaskableBase(ModelBase):
+
     __abstract__ = True
 
     blockchain_task_uuid = db.Column(db.String)
 
-    @hybrid_property
-    def blockchain_status(self):
-        if self.blockchain_task_uuid:
-            task = bt.get_blockchain_task(self.blockchain_task_uuid)
-
-            return task.get('status', 'ERROR')
-        else:
-            return 'UNKNOWN'
+    # Present status, and time of last update (according to worker) to ensure the present blockchain_status 
+    # is the newest (since order of ack's is not guaranteed)
+    blockchain_status   = db.Column(db.Enum(BlockchainStatus), default=BlockchainStatus.PENDING)
+    blockchain_hash = db.Column(db.String)
+    last_worker_update = db.Column(db.DateTime)
+    @declared_attr
+    def messages(cls):
+        return db.relationship('WorkerMessages', primaryjoin=lambda: db.foreign(WorkerMessages.blockchain_task_uuid)==cls.blockchain_task_uuid, lazy=True)
 
 
 class SoftDelete(object):