--- a/app/server/schemas.py
+++ b/app/server/schemas.py
@@ -30,7 +30,7 @@ class SchemaBase(Schema):
 class BlockchainTaskableSchemaBase(SchemaBase):
 
     blockchain_task_uuid  = fields.Str(dump_only=True)
-    #blockchain_status   = fields.Function(lambda obj: obj.blockchain_status.name)
+    blockchain_status   = fields.Function(lambda obj: obj.blockchain_status.name)
 
 class UserSchema(SchemaBase):
     first_name              = fields.Str()