While working on multi-org support, I noticed that `filter_by_org` was being run an inordinate amount of times for the credit_transfer endpoint, indicating that many queries were being run for what should be a relatively simple API call. After some investigation, I traced the issue to how we use schemas. As an example, when this line is run: `lat = fields.Function(lambda obj: obj.recipient_transfer_account.primary_user.lat)`, it actually queried `recipient_transfer_account` separately for _each and every row_. The superfluous query happened for for `Exchange`, `UploadedResource`, and twice for `TransferAccount` (sender and recipient).  This PR remedies it by changing our lazy strategy to `joined` for those rows. 
  
I ran a small benchmark locally comparing the original code to the changes in this PR. I only ran these locally, not on dev, so factor in that this is based on a database and app running on a machine with ample RAM and an SSD. I imagine the performance gulf will be larger on our smaller AWS instances.
`/api/v1/credit_transfer/?get_stats=true&per_page=###&page=1&org=2 `    
Original code:
```
per_page = 100
1712ms for GET request
417 total DB queries

per_page=1000
8100ms for GET request
2018 total DB queries
```  
Accelerated code:
```
per_page = 100
242ms for GET request
19 total DB queries

per_page=1000
1146ms for GET request
19 total DB queries
```  
TL;DR: When run locally these changes accelerate the `/credit_transfer` endpoint **by a factor of 7**